+++ core/src/ru/maklas/genetics/utils/functions/LinearFunction.java
@@ -2,16 +2,16 @@ package ru.maklas.genetics.utils.functions;
 
 public class LinearFunction implements GraphFunction {
 
-    public float k;
-    public float b;
+    public double k;
+    public double b;
 
-    public LinearFunction(float k, float b) {
+    public LinearFunction(double k, double b) {
         this.k = k;
         this.b = b;
     }
 
     @Override
-    public float f(float x) {
+    public double f(double x) {
         return k * x + b;
     }
 }
