+++ core/src/ru/maklas/genetics/utils/functions/OnlyPositiveFunction.java
@@ -2,21 +2,20 @@ package ru.maklas.genetics.utils.functions;
 
 public class OnlyPositiveFunction implements GraphFunction {
 
-
     public GraphFunction delegate;
-    float negativeY = 0;
+    double negativeY = 0;
 
     public OnlyPositiveFunction(GraphFunction delegate) {
         this.delegate = delegate;
     }
 
-    public OnlyPositiveFunction(GraphFunction delegate, float negativeY) {
+    public OnlyPositiveFunction(GraphFunction delegate, double negativeY) {
         this.delegate = delegate;
         this.negativeY = negativeY;
     }
 
     @Override
-    public float f(float x) {
+    public double f(double x) {
         return x < 0 ? negativeY : delegate.f(x);
     }
 }
